fix: producing incorrect types for commonjs root export
Fix using the package producing a `This expression is not callable.` TypeScript error. This is due to the used version of unbuild (see https://github.com/unjs/unbuild/issues/457) and was fixed in the meantime meaning that simply updating unbuild is enough to fix the issue. (Note, that 3.5.0 is the latest version of unbuild that can be installed without also updating TypeScript.)